+++ src/main/kotlin/com/github/vitallium/rubylsp/settings/DisplayName.kt
@@ -1,0 +1,4 @@
+package com.github.vitallium.rubylsp.settings
+
+@Target(AnnotationTarget.FIELD)
+annotation class DisplayName(val name: String)